pyproject.toml
[project]
name = "hermes"
version = "1.0.0"
description = "Crowdin i18n translation management tool with TUI and CLI support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.89.1",
    "requests>=2.32.0",
    "google-genai>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
hermes = "hermes.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hermes"]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "ARG001",  # unused function argument (common in callbacks)
    "ARG002",  # unused method argument
]

[tool.ruff.lint.isort]
known-first-party = ["hermes"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py are fine
"tests/*" = ["ARG", "PLR"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyinstaller]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
    "ruff>=0.14.11",
]
# PyInstaller configuration for single executable
# Run: pyinstaller hermes.spec
